backend/coatstore/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==3.14.0
djangorestframework==3.14.0
psycopg2-binary==2.9.7
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
django-filter==23.2
drf-yasg